http: move some public API bits to its final places and mark them as `@ApiMayChange`
