+++ akka-http-core/src/main/scala/akka/http/scaladsl/model/http2/Http2Exception.scala
@@ -4,6 +4,7 @@
 
 package akka.http.scaladsl.model.http2
 
+import akka.annotation.ApiMayChange
 import akka.http.impl.engine.http2.Http2Protocol.ErrorCode
 
 import scala.util.control.NoStackTrace
@@ -11,11 +12,13 @@ import scala.util.control.NoStackTrace
 /**
  * Base class for HTTP2 exceptions.
  */
+@ApiMayChange
 class Http2Exception(msg: String) extends RuntimeException(msg)
 
 /**
  * Exception that will be reported on the request entity stream when the peer closed the stream.
  */
+@ApiMayChange
 class PeerClosedStreamException(val streamId: Int, val errorCode: String, val numericErrorCode: Int)
   extends Http2Exception(f"Stream with ID [$streamId%d] was closed by peer with code $errorCode%s(0x$numericErrorCode%02x)") with NoStackTrace {
   private[http] def this(streamId: Int, errorCode: ErrorCode) = this(streamId, errorCode.toString, errorCode.id)
